Small fixes v5
