+++ BookMart/app/core/App.php
@@ -61,9 +61,7 @@ class App{
         ],
         'bookSeller' => [
             'BookSellerHome' => ['index','storePage', 'viewSales'],
-            'BookSellerSales' => ['index'], 
-            'BookSellerListings' => ['index', 'addBook', 'updateBook', 'deleteBook'], 
-            'BookSellerController' => ['index', 'orders','trackOrder','register','reviews','addReview','submitReview','myProfile','updateLoginDetails','updatePersonalDetails', 'updateBankDetails', 'uploadProfilePicture'],
+            'BookSellerController' => ['index', 'listings', 'addBook', 'updateBook', 'deleteBook', 'orders','trackOrder','register','reviews','addReview','submitReview','myProfile','updateLoginDetails','updatePersonalDetails', 'updateBankDetails', 'uploadProfilePicture'],
             'Articles' => ['index', 'create', 'addArticle', 'detail', 'update', 'updateArticle', 'deleteArticle'],
             'BookSellerSidebar' => ['index'],
             'Auctions' => ['index', 'details', 'createAuction', 'updateBid', 'buyNow', 'completeAuction', 'cancelAuction', 'withdraw'],
