+++ BookMart/app/models/AuctionModel.php
@@ -56,11 +56,11 @@ class AuctionModel {
     } 
 
     public function getAuctionWithBook($id) {
-        $query = "SELECT a.*, b.title, b.author, b.cover_image AS image, b.description, b.book_condition, u.username AS seller_name
-                  FROM auction a
-                  JOIN book b ON a.book_id = b.id
-                  JOIN user u ON a.seller_id = u.ID
-                  WHERE a.id = :id";
+        $query = "SELECT auction.*, book.title, book.author, book.cover_image AS image, book.description, book.book_condition, bookseller.full_name AS seller_name, bookseller.rating AS seller_rating, bookseller.profile_picture AS seller_image
+                  FROM auction
+                  JOIN book ON auction.book_id = book.id
+                  JOIN bookseller ON auction.seller_id = bookseller.user_id
+                  WHERE auction.id = :id";
         return $this->query($query, ['id' => $id])[0] ?? null;
     }
