+++ BookMart/app/views/commonSidebar.view.php
@@ -8,7 +8,7 @@ $currentPath = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
         <ul>
             <li><button class="add-book-bttn"><span class="compose-icon"><i class="fa-solid fa-plus"></i></span>Add book</button></li>
             <li><button class="add-book-bttn" onClick="location.href='<?= ROOT ?>/articles/create';"><span class="compose-icon"><i class="fa-solid fa-plus"></i></span>Create Article</button></li>
-            <li><a href="<?= ROOT ?>/bookSellerListings" class="<?= strpos($currentPath, '/bookSellerListings') !== false ? 'active' : '' ?>" class="active" ><i class="fa-solid fa-book"></i>My Listings</a></li>
+            <li><a href="<?= ROOT ?>/bookSellerController/listings" class="<?= strpos($currentPath, '/bookSellerController/listings') !== false ? 'active' : '' ?>" class="active" ><i class="fa-solid fa-book"></i>My Listings</a></li>
             <li><a href="<?= ROOT ?>/auctions" class="<?= strpos($currentPath, '/auctions') !== false ? 'active' : '' ?>"><i class="fa-solid fa-chart-column"></i>Auctions</a></li>
             <li><a href="<?= ROOT ?>/articles" class="<?= strpos($currentPath, '/articles') !== false && strpos($currentPath, '/myArticles') === false ? 'active' : '' ?>"><i class="fa-solid fa-cart-plus"></i>Articles</a></li>
             <li>
@@ -209,7 +209,7 @@ $currentPath = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
     <div id="add-book-modal"class="modal hidden">
         <div class="modal-overlay"></div>
             <div class="modal-content">
-                <form class="add-book-form" method="POST" action="<?= ROOT ?>/bookSellerListings/addBook" enctype="multipart/form-data">
+                <form class="add-book-form" method="POST" action="<?= ROOT ?>/bookSellerController/addBook" enctype="multipart/form-data">
                     <h2 class="full-width">Add Book</h2>
 
                     <!-- Title -->
