+++ BookMart/public/assets/CSS/notifications.css
@@ -112,47 +112,34 @@ body {
     width: 100%;
 }
 
-.view-notification-btn, .mark-read-btn, .delete-notification-btn {
-    padding: 5px 10px;
-    border: none;
-    border-radius: 4px;
+.notification-actions button{
+    padding: 15px 50px;
     cursor: pointer;
-    font-size: 0.9em;
-    width: 150px;
-    height: 40px;
-}
-
-.view-notification-btn {
-    background-color: #BE5715;
-    color: #ffffff;
-}
-
-.view-notification-btn:hover {
-    background-color: #e7691a;
-    color: #ffffff;
-    transform: translateY(-2px);
+    font-weight: 700;
 }
 
 .mark-read-btn {
-    background-color: #376ea6;
-    color: #ffffff;
+    border:solid 3px #02224E;
+    background-color: #02224e00;
+    color: #02224E;
 }
 
 .mark-read-btn:hover {
-    background-color: #4996e3;
+    border:solid 3px #02224E;
+    background-color: #02224e;
     color: #ffffff;
-    transform: translateY(-2px);
 }
 
 .delete-notification-btn {
-    background-color: #cf0000;
-    color: #ffffff;
+    background-color: #c2413800;
+    border:solid 3px #ff0000;
+    color: #ff0000;
 }
 
 .delete-notification-btn:hover {
     background-color: #ff0000;
+    border:solid 3px #ff0000;
     color: #ffffff;
-    transform: translateY(-2px);
 }
 
 .view-more-button {
